requirements.txt
Django==1.11.7
python-dateutil==2.6.1
requests==2.18.4
bleach==2.1.2